refactor: move info endpoint to the debug api
