+++ pkg/api/api_test.go
@@ -66,7 +66,6 @@ type testServerOptions struct {
 	WsPath             string
 	Tags               *tags.Tags
 	GatewayMode        bool
-	BeeMode            api.BeeNodeMode
 	WsPingPeriod       time.Duration
 	Logger             logging.Logger
 	PreventRedirect    bool
@@ -102,7 +101,6 @@ func newTestServer(t *testing.T, o testServerOptions) (*http.Client, *websocket.
 	}
 	s := api.New(o.Tags, o.Storer, o.Resolver, o.Pss, o.Traversal, o.Pinning, o.Feeds, o.Post, o.PostageContract, o.Steward, signer, o.Authenticator, o.Logger, nil, api.Options{
 		CORSAllowedOrigins: o.CORSAllowedOrigins,
-		BeeMode:            o.BeeMode,
 		GatewayMode:        o.GatewayMode,
 		WsPingPeriod:       o.WsPingPeriod,
 		Restricted:         o.Restricted,
